[clang-tidy] altera-id-dependent-backward-branch: test all inferred assignments
There are six cases of inferrence in the code, test all of them:
a variable initialized from a variable/field, a variable assigned from a
variable/field, a field assigned from a variable/field.

It can be now seen that notes point to variable/field declarations, not actual
assignment statements.