[clang-tidy] altera-id-dependent-backward-branch: find more function calls
Previously 'hasDescendant' was used to look for direct function calls, but it
looks among children only and never matches the root node. Tests passed before
because there was a node for an implicit conversion from `size_t` to `int`. Now
we look both among children and try to match the node itself. Tests are updated
to have both cases.